Order line check update

+++ CHubMVC/CHubBLL/G_NETAVL_BLL.cs
@@ -22,7 +22,7 @@ namespace CHubBLL
             dal = new G_NETAVL_DAL(db);
         }
 
-        public int GetSpecifyNETAVL(string sysId, string partNo, string wareHouse)
+        public decimal GetSpecifyNETAVL(string sysId, string partNo, string wareHouse)
         {
             return dal.GetSpecifyNETAVL(sysId, partNo, wareHouse);
         }
